Update index.html
Changed the structure so Certification bodies is at bottom. When selecting Presses and then selecting or unselecting hydraulic or pneumatic you will see the additional attributes highlighting. Also when selecting the origin and destination markets you will see the correct certification getting checked, i.e. Select US to Italy and the CE Marking is selected. I have also fixed the order in which the output is given so standards are listed together first, followed by directives, and then certification requirements